Implement an OrderPizza UseCase
This commit demonstrates TDD, DI, and the Single Responsibility Principle.

It's hard to tell afterwards, but I test drove the order pizza use case.
It is only responsible for validating an order and dispatching it to the
appropriate upstream services. It doesn't know HOW to make a pizza, but
it knows how to collaborate through a pizza repository that is able to
encapsulate the knowledge of how to send a pizza order to an upstream service.