Use cases should not depend on implementations
Rather, these interfaces should be declared where they will be consumed.
This results in double-entry bookkeeping, because interfaces will need
to be changed on both sides when a change is necessary, but it prevents
your packages from becoming explitly tied up in one another.

In this case, for a clean architecture app, it makes no sense for the
use cases that encode the actual business rules of the system to depend
on anything from an implementation (even one that is a generic interface).