Adds a minimal acceptance test with gexec
This exposed the fact that we wrote from the request body multiple
times in our http handler. This ... never worked.

Moving the responsibility for reading the entire request body into our
param reader was a nice little win that came out of this because our
http handler was previously doing too much work.